refactor(theme): relocate react-components scaffold under components
