+++ src/components/scaffold/election.tsx
